refactor: Email 전송 구조 리펙토링
