+++ src/main/java/org/codequistify/master/application/authentication/dto/LogInRequest.java
@@ -1,0 +1,9 @@
+package org.codequistify.master.application.authentication.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LogInRequest(
+        @NotBlank String email,
+        @NotBlank String password
+) {
+}
