+++ src/main/java/org/codequistify/master/application/authentication/dto/SocialLogInRequest.java
@@ -1,0 +1,8 @@
+package org.codequistify.master.application.authentication.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record SocialLogInRequest(
+        @NotBlank String code
+) {
+}
