+++ src/main/java/org/codequistify/master/application/authentication/vo/OAuthData.java
@@ -1,0 +1,10 @@
+package org.codequistify.master.application.authentication.vo;
+
+public record OAuthData(
+        OAuthToken token,
+        OAuthResource resource
+) {
+    public static OAuthData of(OAuthToken token, OAuthResource resource) {
+        return new OAuthData(token, resource);
+    }
+}
