+++ src/main/java/org/codequistify/master/application/authentication/vo/OAuthResource.java
@@ -1,0 +1,8 @@
+package org.codequistify.master.application.authentication.vo;
+
+public record OAuthResource(
+        String id,
+        String email,
+        String name
+) {
+}
