+++ src/main/java/org/codequistify/master/application/authentication/vo/ResourceOfKakao.java
@@ -1,0 +1,8 @@
+package org.codequistify.master.application.authentication.vo;
+
+import java.util.Map;
+
+public record ResourceOfKakao(
+        Map<String, String> properties
+) {
+}
