--- src/main/java/org/codequistify/master/core/domain/authentication/dto/LogInRequest.java
@@ -1,9 +1,0 @@
-package org.codequistify.master.core.domain.authentication.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record LogInRequest(
-        @NotBlank String email,
-        @NotBlank String password
-) {
-}
