--- src/main/java/org/codequistify/master/core/domain/authentication/dto/SocialLogInRequest.java
@@ -1,8 +1,0 @@
-package org.codequistify.master.core.domain.authentication.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record SocialLogInRequest(
-        @NotBlank String code
-) {
-}
