--- src/main/java/org/codequistify/master/core/domain/authentication/vo/OAuthData.java
@@ -1,10 +1,0 @@
-package org.codequistify.master.core.domain.authentication.vo;
-
-public record OAuthData(
-        OAuthToken token,
-        OAuthResource resource
-) {
-    public static OAuthData of(OAuthToken token, OAuthResource resource) {
-        return new OAuthData(token, resource);
-    }
-}
