--- src/main/java/org/codequistify/master/core/domain/authentication/vo/OAuthResource.java
@@ -1,8 +1,0 @@
-package org.codequistify.master.core.domain.authentication.vo;
-
-public record OAuthResource(
-        String id,
-        String email,
-        String name
-) {
-}
