--- src/main/java/org/codequistify/master/core/domain/authentication/vo/ResourceOfNaver.java
@@ -1,8 +1,0 @@
-package org.codequistify.master.core.domain.authentication.vo;
-
-import java.util.Map;
-
-public record ResourceOfNaver(
-        Map<String, String> response
-) {
-}
